+++ src/render/widget.rs
@@ -18,6 +18,7 @@ use sdl2::render::Canvas;
 use sdl2::video::Window;
 
 use crate::render::callbacks::*;
+use crate::render::canvas_helper::CanvasHelper;
 use crate::render::layout_cache::LayoutContainer;
 use crate::render::widget_cache::WidgetContainer;
 use crate::render::widget_config::*;
@@ -336,6 +337,8 @@ impl BaseWidget {
     }
 }
 
+impl CanvasHelper for BaseWidget {}
+
 /// Implementation for drawing a `BaseWidget`, with the `Widget` trait objects applied.
 impl Widget for BaseWidget {
     fn draw(&mut self, mut _canvas: &mut Canvas<Window>) {
@@ -346,20 +349,8 @@ impl Widget for BaseWidget {
 
         _canvas.fill_rect(self.get_drawing_area()).unwrap();
 
-        if self.get_config().get_numeric(CONFIG_BORDER_WIDTH) > 0 && base_color != border_color {
-            _canvas.set_draw_color(border_color);
-
-            for border in 0..self.get_config().get_numeric(CONFIG_BORDER_WIDTH) {
-                _canvas
-                    .draw_rect(Rect::new(
-                        self.config.to_x(border),
-                        self.config.to_y(border),
-                        self.get_config().get_size(CONFIG_SIZE)[0] - (border as u32 * 2),
-                        self.get_config().get_size(CONFIG_SIZE)[1] - (border as u32 * 2),
-                    ))
-                    .unwrap();
-            }
-        }
+        _canvas.set_draw_color(border_color);
+        self.draw_bounding_box(_canvas);
     }
 
     default_widget_functions!();